Backend/requirements.txt
Flask
flask-api
flask_cors
pyrebase4
python-dotenv
pytz